fix: Fix literal string allocation issue
In Jack OS literal strings are poorly handled. When calling functions like Output.printstring("example") "example" would be allocated space in the heap but there is no pointer to this memory location and this string is not deallocated. This issue continues to persist when printing ints to the screen but this fix allows for +30k score (when playing semi optimal atleast)